% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringOperatingSet.R
\name{Excel2sqlstring}
\alias{Excel2sqlstring}
\title{将向量转化为SQL-where的械}
\usage{
Excel2sqlstring(data)
}
\arguments{
\item{data}{数据}
}
\value{
返回值
}
\description{
将向量转化为SQL-where的械
}
\examples{
Excel2sqlstring('aa');
}
