% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringOperatingSet.R
\name{getDataFromExcel}
\alias{getDataFromExcel}
\title{从Excel中读取数据}
\usage{
getDataFromExcel(file, sheet = 1)
}
\arguments{
\item{file}{文件名}

\item{sheet}{页签名}
}
\value{
返回值
}
\description{
从Excel中读取数据
}
\examples{
getDataFromExcel('aa.xlsx',1)
}
