% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringOperatingSet.R
\name{writeDataToExcel}
\alias{writeDataToExcel}
\title{将数据写入Excel}
\usage{
writeDataToExcel(data, fileName, sheetName)
}
\arguments{
\item{data}{数据}

\item{fileName}{文件名}

\item{sheetName}{页签名}
}
\value{
返回值
}
\description{
将数据写入Excel
}
\examples{
writeDataToExcel（letters,'samle.xlsx','sheet1'）
}
